fix(config): respect rate limit env vars when config file is present
When a config file was present but did not contain a rate_limit
section, the default values for rate limiting would be applied and
override any environment variables that were set.

This commit fixes the issue by only applying the rate limit settings
from the config file if the rate_limit section is explicitly present in
the file.

This also fixes a type mismatch for the RPS value in the file config.

+++ internal/config/file_test.go
@@ -120,7 +120,7 @@ http:
 		t.Error("expected rate_limit.enabled true")
 	}
 	if cfg.HTTP.RateLimit.RPS != 50 {
-		t.Errorf("expected rate_limit.rps 50, got %d", cfg.HTTP.RateLimit.RPS)
+		t.Errorf("expected rate_limit.rps 50, got %f", cfg.HTTP.RateLimit.RPS)
 	}
 }
 
@@ -344,7 +344,7 @@ func TestMinimalConfigDefaults(t *testing.T) {
 		t.Errorf("expected HTTPPort %d, got %d", DefaultHTTPPort, cfg.HTTPPort)
 	}
 	if cfg.RateLimitRPS != float64(DefaultRateLimitRPS) {
-		t.Errorf("expected RateLimitRPS %d, got %f", DefaultRateLimitRPS, cfg.RateLimitRPS)
+		t.Errorf("expected RateLimitRPS %f, got %f", float64(DefaultRateLimitRPS), cfg.RateLimitRPS)
 	}
 	if cfg.RateLimitBurst != DefaultRateLimitBurst {
 		t.Errorf("expected RateLimitBurst %d, got %d", DefaultRateLimitBurst, cfg.RateLimitBurst)
